update apache plugin

+++ source/apache/usr/local/emhttp/plugins/apache/include/ClearLogFile.php
@@ -1,4 +1,9 @@
 <?
-$log = $_GET['log'];
-file_put_contents($log, '');
+$base = '/var/log/httpd/';
+$log  = realpath($_POST['log']);
+
+//check that log file is in the base path and exists
+if(!strpos($log, $base) && file_exists($log)){
+    file_put_contents($log, '');
+}
 ?>
